[hotfix] Rename flink-ml-api to flink-ml-core
