+++ flink-ml-core/src/main/java/org/apache/flink/ml/api/Transformer.java
@@ -16,24 +16,17 @@
  * limitations under the License.
  */
 
-package org.apache.flink.ml.api.core;
+package org.apache.flink.ml.api;
 
 import org.apache.flink.annotation.PublicEvolving;
-import org.apache.flink.table.api.Table;
 
 /**
- * Estimators are responsible for training and generating Models.
+ * A Transformer is an AlgoOperator with the semantic difference that it encodes the Transformation
+ * logic, such that a record in the output typically corresponds to one record in the input. In
+ * contrast, an AlgoOperator is a better fit to express aggregation logic where a record in the
+ * output could be computed from an arbitrary number of records in the input.
  *
- * @param <E> class type of the Estimator implementation itself.
- * @param <M> class type of the Model this Estimator produces.
+ * @param <T> The class type of the Transformer implementation itself.
  */
 @PublicEvolving
-public interface Estimator<E extends Estimator<E, M>, M extends Model<M>> extends Stage<E> {
-    /**
-     * Trains on the given inputs and produces a Model.
-     *
-     * @param inputs a list of tables
-     * @return a Model
-     */
-    M fit(Table... inputs);
-}
+public interface Transformer<T extends Transformer<T>> extends AlgoOperator<T> {}
