+++ flink-ml-core/src/main/java/org/apache/flink/ml/builder/PipelineModel.java
@@ -16,10 +16,13 @@
  * limitations under the License.
  */
 
-package org.apache.flink.ml.api.core;
+package org.apache.flink.ml.builder;
 
 import org.apache.flink.annotation.PublicEvolving;
 import org.apache.flink.annotation.VisibleForTesting;
+import org.apache.flink.ml.api.AlgoOperator;
+import org.apache.flink.ml.api.Model;
+import org.apache.flink.ml.api.Stage;
 import org.apache.flink.ml.param.Param;
 import org.apache.flink.ml.util.ParamUtils;
 import org.apache.flink.ml.util.ReadWriteUtils;
