+++ flink-ml-core/src/main/java/org/apache/flink/ml/distance/DistanceMeasure.java
