+++ flink-ml-core/src/main/java/org/apache/flink/ml/distance/EuclideanDistanceMeasure.java
