+++ flink-ml-core/src/main/java/org/apache/flink/ml/param/DoubleArrayParam.java
