+++ flink-ml-lib/pom.xml
@@ -32,7 +32,7 @@ under the License.
   <dependencies>
     <dependency>
       <groupId>org.apache.flink</groupId>
-      <artifactId>flink-ml-api</artifactId>
+      <artifactId>flink-ml-core</artifactId>
       <version>${project.version}</version>
       <scope>provided</scope>
     </dependency>
