+++ flink-ml-uber/pom.xml
@@ -37,7 +37,7 @@ under the License.
   <dependencies>
     <dependency>
       <groupId>org.apache.flink</groupId>
-      <artifactId>flink-ml-api</artifactId>
+      <artifactId>flink-ml-core</artifactId>
       <version>${project.version}</version>
     </dependency>
     <dependency>
@@ -62,7 +62,7 @@ under the License.
             <configuration>
               <artifactSet>
                 <includes combine.children="append">
-                  <include>org.apache.flink:flink-ml-api</include>
+                  <include>org.apache.flink:flink-ml-core</include>
                   <include>org.apache.flink:flink-ml-lib_${scala.binary.version}</include>
                   <include>com.github.fommil.netlib:core</include>
                 </includes>
